Move EntryInfosForPatient to record
This was a very large move because of the certain global maps
made in the exporter package. Most of cat1_view_helper.go and
hqmf_template_resolver.go had to be moved to the models package
and was dumped into the record.go file for simplicity. The next
stage of this refactor will be moving this logic out of models
and into its own package that will then act as a service for any
library to use for Health Data Standard specific look ups,
transformations, etc.
